Fixing jankiness in keyboard (still noticable)
