+++ finance/docs/UI_PATTERNS_AND_BEST_PRACTICES.md
@@ -248,4 +248,86 @@ Container(
 ).tappable(
   onTap: _openSettings,
 );
+```
+
+### Keyboard-Aware Bottom Sheets (Post-Refactor Pattern)
+
+The app now includes a refined, high-performance approach to bottom sheets that automatically handle keyboard interactions without jank. This pattern is built into `BottomSheetService` and leverages the internal `_KeyboardAwareBottomSheet` widget.
+
+#### Best Practice: Using Bottom Sheets with Text Input
+
+```dart
+// ✅ Recommended: Use popupWithKeyboard for sheets with text fields
+void _showTransactionForm(BuildContext context) {
+  BottomSheetService.showCustomBottomSheet(
+    context,
+    Column(
+      mainAxisSize: MainAxisSize.min,
+      children: [
+        TextInput(hintText: "Transaction title"),
+        TextInput(hintText: "Amount", keyboardType: TextInputType.number),
+        TextInput(hintText: "Notes", maxLines: 3),
+      ],
+    ),
+    title: "Add Transaction",
+    popupWithKeyboard: true,     // Optimizes snap sizes for keyboard
+    resizeForKeyboard: true,     // Enables smooth tracking (default)
+    snapSizes: [0.9, 1.0],       // Keyboard-friendly snap points
+  );
+}
+```
+
+#### Technical Pattern: Real-Time Keyboard Synchronization
+
+The `_KeyboardAwareBottomSheet` widget provides frame-by-frame keyboard tracking:
+
+```dart
+// This pattern is internal to BottomSheetService but demonstrates the approach
+class _KeyboardAwareBottomSheet extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    // Real-time animated values from MediaQuery
+    final keyboardHeight = MediaQuery.viewInsetsOf(context).bottom;
+    
+    // Simple padding approach - no Stack complexity
+    return Padding(
+      padding: EdgeInsets.only(bottom: keyboardHeight),
+      child: child,
+    );
+  }
+}
+```
+
+**Key Benefits:**
+- **Zero Jank:** Content moves smoothly with keyboard animation
+- **Simple Architecture:** No complex timing logic or competing animations
+- **Automatic:** Works for all bottom sheet types (draggable, standard, options)
+- **Backward Compatible:** Existing code continues to work unchanged
+
+#### Bottom Sheet Performance Guidelines
+
+```dart
+// ✅ Good: Use appropriate snap sizes for content type
+BottomSheetService.showCustomBottomSheet(
+  context,
+  content,
+  snapSizes: [0.3, 0.6, 0.9],    // Multiple options for scrollable content
+  resizeForKeyboard: true,        // Enable for forms
+);
+
+// ✅ Good: Optimize for keyboard scenarios
+BottomSheetService.showCustomBottomSheet(
+  context,
+  formContent,
+  popupWithKeyboard: true,        // Uses [0.9, 1.0] snap sizes
+  resizeForKeyboard: true,        // Smooth keyboard tracking
+);
+
+// ✅ Good: Disable for static content that doesn't need keyboard
+BottomSheetService.showOptionsBottomSheet(
+  context,
+  title: "Choose action",
+  options: menuOptions,
+  resizeForKeyboard: false,       // Static options don't need keyboard tracking
+);
 ```
