+++ finance/lib/shared/widgets/animations/tappable_widget.dart
@@ -6,7 +6,6 @@ import 'animation_utils.dart';
 import 'faded_button.dart';
 import '../../../core/services/platform_service.dart';
 import '../../../core/services/animation_performance_service.dart';
-import '../../utils/performance_optimization.dart';
 
 /// A widget that provides customizable tap feedback with animations
 /// Phase 6.2 implementation with performance optimization
@@ -69,12 +68,7 @@ class _TappableWidgetState extends State<TappableWidget>
     _isAndroid = platform == PlatformOS.isAndroid;
     _isDesktop = PlatformService.isDesktop;
     
-    // Track platform optimization usage
-    PerformanceOptimizations.trackPlatformOptimization(
-      'TappableWidget', 
-      platform.toString(), 
-      'cached platform detection'
-    );
+    // Platform optimization tracking removed
 
     _controller = AnimationUtils.createController(
       vsync: this,
@@ -178,11 +172,7 @@ class _TappableWidgetState extends State<TappableWidget>
   Widget build(BuildContext context) {
     // Phase 3: Use cached platform detection for better performance
     if (_isIOS) {
-      PerformanceOptimizations.trackPlatformOptimization(
-        'TappableWidget', 
-        'iOS', 
-        'FadedButton'
-      );
+      // Platform optimization tracking removed
       // Disable the button when no interaction callbacks are provided
       final bool _isDisabled =
           widget.onTap == null && widget.onLongPress == null && widget.onDoubleTap == null;
@@ -273,11 +263,7 @@ class _TappableWidgetState extends State<TappableWidget>
         widget.onDoubleTap != null) {
       
       // Use GestureDetector for all platforms (no splash effects)
-      PerformanceOptimizations.trackPlatformOptimization(
-        'TappableWidget', 
-        'All Platforms', 
-        'GestureDetector without splash'
-      );
+      // Platform optimization tracking removed
       gestureChild = GestureDetector(
         onTapDown: widget.animationType != TapAnimationType.none
             ? _handleTapDown
